separated pricing page content from components
